+++ src/components/ui/ConfirmModal/ConfirmModal.tsx
@@ -7,7 +7,7 @@ interface ConfirmModalProps {
     onClose: () => void;
     onConfirm: () => void;
     title: string;
-    message: React.ReactNode;
+    toDelete: React.ReactNode;
     confirmLabel?: string;
 }
 
@@ -16,11 +16,13 @@ export const ConfirmModal: React.FC<ConfirmModalProps> = ({
     onClose,
     onConfirm,
     title,
-    message,
+    toDelete: message,
     confirmLabel = "Delete Permanently",
 }) => (
-    <Modal isOpen={isOpen} onClose={onClose} title={title}>
-        <p>{message}</p>
+    <Modal isOpen={isOpen} onClose={onClose} title={`Delete ${title}`}>
+        <p>
+            Are you sure you want to delete <strong>{message}</strong>?
+        </p>
         <button onClick={onConfirm} className={styles.submitButton}>
             {confirmLabel}
         </button>